refactor packaging and naming
