+++ src/main/kotlin/com/crafto/crafto_backend/controller/AuthorizationController.kt
@@ -1,7 +1,7 @@
 package com.crafto.crafto_backend.controller
 
-import com.crafto.crafto_backend.dto.AuthorizationRequest
-import com.crafto.crafto_backend.dto.AuthorizationResponse
+import com.crafto.crafto_backend.dto.request.AuthorizationRequest
+import com.crafto.crafto_backend.dto.response.AuthorizationResponse
 import com.crafto.crafto_backend.service.AuthenticationService
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PostMapping
