--- src/main/kotlin/com/crafto/crafto_backend/dto/ErrorResponse.kt
@@ -1,9 +1,0 @@
-package com.crafto.crafto_backend.dto
-
-import java.time.Instant
-
-data class ErrorResponse(
-    val code: String,
-    val message: String,
-    val timestamp: Instant = Instant.now()
-)
