+++ src/main/kotlin/com/crafto/crafto_backend/dto/response/AuthorizationResponse.kt
@@ -1,4 +1,4 @@
-package com.crafto.crafto_backend.dto
+package com.crafto.crafto_backend.dto.response
 
 data class AuthorizationResponse(
     val userId: String,
