+++ src/main/kotlin/com/crafto/crafto_backend/dto/response/CategoryResponse.kt
@@ -1,4 +1,4 @@
-package com.crafto.crafto_backend.dto
+package com.crafto.crafto_backend.dto.response
 
 data class CategoryResponse(
     val categoryId:String,
