+++ src/main/kotlin/com/crafto/crafto_backend/dto/response/CustomerIssueDetailsResponse.kt
@@ -1,4 +1,4 @@
-package com.crafto.crafto_backend.dto
+package com.crafto.crafto_backend.dto.response
 
 import com.crafto.crafto_backend.database.entity.CustomerIssueStatus
