+++ src/main/kotlin/com/crafto/crafto_backend/dto/response/GovernorateResponse.kt
@@ -1,4 +1,4 @@
-package com.crafto.crafto_backend.dto
+package com.crafto.crafto_backend.dto.response
 
 data class GovernorateResponse(
     val id: String,
