--- src/main/kotlin/com/crafto/crafto_backend/exception/GlobalExceptionHandler.kt
@@ -1,54 +1,0 @@
-package com.crafto.crafto_backend.exception
-
-import com.crafto.crafto_backend.dto.ErrorResponse
-import org.springframework.http.HttpStatus
-import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.MethodArgumentNotValidException
-import org.springframework.web.bind.annotation.ExceptionHandler
-import org.springframework.web.bind.annotation.RestControllerAdvice
-
-@RestControllerAdvice
-class GlobalExceptionHandler {
-
-    @ExceptionHandler(ConflictException::class)
-    fun handleConflict(ex: ConflictException): ResponseEntity<ErrorResponse> {
-        return ResponseEntity.status(HttpStatus.CONFLICT)
-            .body(ErrorResponse("CONFLICT", ex.message ?: "Conflict occurred"))
-    }
-
-    @ExceptionHandler(NotFoundException::class)
-    fun handleNotFound(ex: NotFoundException): ResponseEntity<ErrorResponse> {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND)
-            .body(ErrorResponse("NOT_FOUND", ex.message ?: "Resource not found"))
-    }
-
-    @ExceptionHandler(BadRequestException::class)
-    fun handleBadRequest(ex: BadRequestException): ResponseEntity<ErrorResponse> {
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-            .body(ErrorResponse("BAD_REQUEST", ex.message ?: "Bad request"))
-    }
-
-    @ExceptionHandler(ForbiddenException::class)
-    fun handleForbidden(ex: ForbiddenException): ResponseEntity<ErrorResponse> {
-        return ResponseEntity.status(HttpStatus.FORBIDDEN)
-            .body(ErrorResponse("FORBIDDEN", ex.message ?: "Access denied"))
-    }
-
-    @ExceptionHandler(MethodArgumentNotValidException::class)
-    fun handleValidationExceptions(
-        ex: MethodArgumentNotValidException
-    ): ResponseEntity<Map<String, Any>> {
-        val errors = ex.bindingResult.fieldErrors.associate {
-            it.field to (it.defaultMessage ?: "Invalid value")
-        }
-
-        return ResponseEntity.badRequest().body(
-            mapOf(
-                "status" to "error",
-                "message" to "Validation failed",
-                "errors" to errors
-            )
-        )
-    }
-}
-
