+++ src/main/kotlin/com/crafto/crafto_backend/mapper/CategoryMapper.kt
@@ -1,7 +1,7 @@
 package com.crafto.crafto_backend.mapper
 
 import com.crafto.crafto_backend.database.entity.Category
-import com.crafto.crafto_backend.dto.CategoryResponse
+import com.crafto.crafto_backend.dto.response.CategoryResponse
 
 fun Category.toCategoryResponse(): CategoryResponse {
     return CategoryResponse(
