+++ src/main/kotlin/com/crafto/crafto_backend/mapper/CraftsmanOfferMapper.kt
@@ -1,8 +1,8 @@
 package com.crafto.crafto_backend.mapper
 
 import com.crafto.crafto_backend.database.entity.CraftsmanOffer
-import com.crafto.crafto_backend.dto.CraftsmanOfferRequest
-import com.crafto.crafto_backend.dto.CraftsmanOfferResponse
+import com.crafto.crafto_backend.dto.request.CraftsmanOfferRequest
+import com.crafto.crafto_backend.dto.response.CraftsmanOfferResponse
 import java.time.Instant
 
 fun CraftsmanOfferRequest.toEntity(isSelected: Boolean) = CraftsmanOffer(
