+++ src/main/kotlin/com/crafto/crafto_backend/mapper/District.kt
@@ -1,7 +1,7 @@
 package com.crafto.crafto_backend.mapper
 
 import com.crafto.crafto_backend.database.entity.District
-import com.crafto.crafto_backend.dto.DistrictResponse
+import com.crafto.crafto_backend.dto.response.DistrictResponse
 
 fun District.toAreaResponse(): DistrictResponse {
     return DistrictResponse(
