+++ src/main/kotlin/com/crafto/crafto_backend/mapper/Governorate.kt
@@ -1,7 +1,7 @@
 package com.crafto.crafto_backend.mapper
 
 import com.crafto.crafto_backend.database.entity.Governorates
-import com.crafto.crafto_backend.dto.GovernorateResponse
+import com.crafto.crafto_backend.dto.response.GovernorateResponse
 
 fun Governorates.toGovernorateResponse(): GovernorateResponse {
     return GovernorateResponse(
