+++ src/main/kotlin/com/crafto/crafto_backend/service/CategoriesService.kt
@@ -2,7 +2,7 @@ package com.crafto.crafto_backend.service
 
 import com.crafto.crafto_backend.mapper.toCategoryResponse
 import com.crafto.crafto_backend.database.repository.CategoryRepository
-import com.crafto.crafto_backend.dto.CategoryResponse
+import com.crafto.crafto_backend.dto.response.CategoryResponse
 import com.crafto.crafto_backend.database.entity.Category
 import org.springframework.data.mongodb.core.query.Query
 import org.springframework.stereotype.Service
@@ -11,7 +11,7 @@ import org.springframework.data.mongodb.core.query.Criteria
 
 
 @Service
-class SharedService (
+class CategoriesService (
     private val repository: CategoryRepository,
     private val mongoTemplate: MongoTemplate
 ){
