+++ src/main/kotlin/com/crafto/crafto_backend/service/CraftsmanService.kt
@@ -12,14 +12,14 @@ import com.crafto.crafto_backend.mapper.toResponse
 import com.crafto.crafto_backend.database.repository.CraftsmanOfferRepository
 import com.crafto.crafto_backend.database.repository.CraftsmanRepository
 import com.crafto.crafto_backend.database.repository.CustomerIssueRepository
-import com.crafto.crafto_backend.dto.CraftsmanOfferRequest
-import com.crafto.crafto_backend.dto.CraftsmanSetupRequest
-import com.crafto.crafto_backend.dto.CraftsmanStatusResponse
+import com.crafto.crafto_backend.dto.request.CraftsmanOfferRequest
+import com.crafto.crafto_backend.dto.request.CraftsmanSetupRequest
+import com.crafto.crafto_backend.dto.response.CraftsmanStatusResponse
 import com.crafto.crafto_backend.exception.BadRequestException
 import com.crafto.crafto_backend.exception.ConflictException
 import com.crafto.crafto_backend.exception.ForbiddenException
 import com.crafto.crafto_backend.exception.NotFoundException
-import com.crafto.crafto_backend.dto.CraftsmanOfferResponse
+import com.crafto.crafto_backend.dto.response.CraftsmanOfferResponse
 import com.crafto.crafto_backend.utils.getFileExtension
 import com.crafto.crafto_backend.utils.validateImageFile
 import org.bson.types.ObjectId
