+++ src/main/kotlin/com/crafto/crafto_backend/service/CustomerService.kt
@@ -11,15 +11,15 @@ import com.crafto.crafto_backend.database.repository.CategoryRepository
 import com.crafto.crafto_backend.database.repository.CraftsmanOfferRepository
 import com.crafto.crafto_backend.database.repository.CustomerIssueRepository
 import com.crafto.crafto_backend.database.repository.CustomerRepository
-import com.crafto.crafto_backend.dto.CustomerIssueRequest
-import com.crafto.crafto_backend.dto.CustomerRequest
-import com.crafto.crafto_backend.dto.CustomerResponse
-import com.crafto.crafto_backend.dto.CustomerSetupRequest
+import com.crafto.crafto_backend.dto.request.CustomerIssueRequest
+import com.crafto.crafto_backend.dto.request.CustomerRequest
+import com.crafto.crafto_backend.dto.response.CustomerResponse
+import com.crafto.crafto_backend.dto.request.CustomerSetupRequest
 import com.crafto.crafto_backend.exception.ConflictException
 import com.crafto.crafto_backend.exception.ForbiddenException
 import com.crafto.crafto_backend.exception.NotFoundException
-import com.crafto.crafto_backend.dto.CustomerIssueDetailsResponse
-import com.crafto.crafto_backend.dto.CustomerIssueResponse
+import com.crafto.crafto_backend.dto.response.CustomerIssueDetailsResponse
+import com.crafto.crafto_backend.dto.response.CustomerIssueResponse
 import com.crafto.crafto_backend.utils.getFileExtension
 import com.crafto.crafto_backend.utils.validateImageFile
 import org.bson.types.ObjectId
